% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxdirectional_to_df.R
\name{boxdirectional_to_df}
\alias{boxdirectional_to_df}
\title{Return testbox directional data from Verifit xml to data frame}
\usage{
boxdirectional_to_df(filename)
}
\arguments{
\item{filename}{An xml file exported from the Verifit}
}
\value{
A data frame containing the testbox directional results.
}
\description{
Takes an input argument of an xml file as exported by the Verifit, and
returns a data frame containing testbox directional results. Data frame is in
long format.
}
\examples{
\dontrun{
df <- boxdirectional_to_df("xxxAbcdE_directionality.xml")
head(df)
}

}
