% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_remdirectional.R
\name{plot_directional}
\alias{plot_directional}
\title{Plot Verifit on-ear or testbox directional results}
\usage{
plot_directional(x, separate.ears = FALSE, main.title = NULL,
  plot.tests = c(1, 2, 3, 4), separate.tests = FALSE,
  test.labs = c("Test 1", "Test 2", "Test 3", "Test 4"))
}
\arguments{
\item{x}{A data frame of Verifit on-ear or testbox directional data, as returned from
\code{XXXdirectional_to_df}}

\item{separate.ears}{An optional boolean value to indicate if ears should be
plotted separately in facets (\code{TRUE}) or averaged in a single plot
(\code{FALSE}).}

\item{main.title}{An optional string that will be used as a main title,
if present.}

\item{plot.tests}{A numeric vector corresponding to which tests should be
included in the plot.  If used with \code{separate.tests = TRUE}, the vector
passed to \code{test.labs} must match \code{plot.tests} in length. Note that
this argument should be numeric, while the data frame returned from
\code{XXXdirectional_to_df} labels tests with a string (e.g., 'test1', test2').}

\item{separate.tests}{An optional boolean value to indicate if test runs should be
plotted separately in facets (\code{TRUE}) or averaged in a single plot
(\code{FALSE}).  This is convenient if multiple directional settings were used
during the test.}

\item{test.labs}{A vector of strings used to label test facets.  Only required
if \code{separate.tests = TRUE}.}
}
\value{
A ggplot2 plot of the data in x.
}
\description{
Takes the data frame returned by \code{XXXdirectional_to_df} and plots the data
using ggplot2.  Data are plotted on a logx scale with color coding for
stimulus level.  The function can be configured to either average across ears
or separate the ears into facets.
}
\examples{
\dontrun{
df <- remdirectional_to_df(speechmap.xml)
plot_remdirectional(df)
}

}
