% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_speechmap.R
\name{plot_speechmap}
\alias{plot_speechmap}
\title{Plot Verifit speechmap results}
\usage{
plot_speechmap(x, separate.ears = FALSE, main.title = NULL)
}
\arguments{
\item{x}{A data frame of Verifit speechmap data, as returned from
\code{speechmap_to_df}}

\item{separate.ears}{An optional boolean value to indicate if ears should be
plotted separately in facets (\code{TRUE}) or averaged in a single plot
(\code{FALSE}).}

\item{main.title}{An optional string that will be used as a main title,
if present.}
}
\value{
A ggplot2 plot of the data in x.
}
\description{
Takes the data frame returned by \code{speechmap_to_df} and plots the data
using ggplot2.  Data are plotted on a logx scale with color coding for
stimulus level.  The function can be configured to either average across ears
or separate the ears into facets.
}
\examples{
\dontrun{
df <- speechmap_to_df(speechmap.xml)
plot_speechmap(df)
}

}
