% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remdirectional_to_df.R
\name{remdirectional_to_df}
\alias{remdirectional_to_df}
\title{Return on-ear directional data from Verifit xml to data frame}
\usage{
remdirectional_to_df(filename)
}
\arguments{
\item{filename}{An xml file exported from the Verifit}
}
\value{
A data frame containing the on-ear directional results.
}
\description{
Takes an input argument of an xml file as exported by the Verifit, and
returns a data frame containing on-ear directional results. Data frame is in
long format.
}
\examples{
\dontrun{
df <- remdirectional_to_df("xxxAbcdE_directionality.xml")
head(df)
}

}
