% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speechmap_to_df.R
\name{speechmap_to_df}
\alias{speechmap_to_df}
\title{Return speech map data from Verifit xml to data frame}
\usage{
speechmap_to_df(filename)
}
\arguments{
\item{filename}{An xml file exported from the Verifit}
}
\value{
A data frame containing the speech map results.
}
\description{
Takes an input argument of an xml file as exported by the Verifit, and
returns a data frame containing speechmap results. Data frame is in long
format.
}
\examples{
\dontrun{
df <- speechmap_to_df("xxxAbcdE_speechmap.xml")
head(df)
}

}
