% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speechmap_to_csv.R
\name{verifit_to_csv}
\alias{verifit_to_csv}
\title{Convert Verifit xml to csv}
\usage{
verifit_to_csv(filename, method = c(speechmap_to_df, remdirectional_to_df,
  boxdirectional_to_df))
}
\arguments{
\item{filename}{An xml file exported from the Verifit}

\item{method}{speechmap_to_df or remdirectional_to_df}
}
\value{
Nothing can be directly returned to the R environment.
}
\description{
Takes input arguments of an xml file as exported by the Verifit and a
'method' (see below), and creates a csv file containing results in the
current working directory. The csv file will share a filename with the input
file name.  The original xml file is preserved.
}
