% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlnumeric.R
\name{xml_to_numeric}
\alias{xml_to_numeric}
\title{Return a numeric vector of data under an xml node}
\usage{
xml_to_numeric(x)
}
\arguments{
\item{x}{A node returned from \code{XML::xmlRoot} that has no children.}
}
\value{
A numeric vector comprising the data in \code{x}
}
\description{
Verifit data is interpreted by R to be a single string.  This function takes
an xml node (returned from \code{XML::xmlRoot}) and converts the content into
a numeric list.
}
\examples{
\dontrun{
xml_to_numeric(xml_data_top[[a]][[b]][[c]])
}

}
